ref: Refactor code to include all files inside Users module
