+++ src/app/app.module.ts
@@ -4,17 +4,15 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
 
-
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MatIconModule } from '@angular/material/icon';
 import { MatTabsModule } from '@angular/material/tabs';
 import { MatCardModule } from '@angular/material/card';
 
 import { AppComponent } from './app.component';
-import { CardComponent } from './components/card/card.component';
-import { MainTabsComponent } from './components/main-tabs/main-tabs.component';
-import { MainTolbarComponent } from './components/main-tolbar/main-tolbar.component';
-
+import { CardComponent } from './users/components/card/card.component';
+import { MainTabsComponent } from './users/components/main-tabs/main-tabs.component';
+import { MainTolbarComponent } from './users/components/main-tolbar/main-tolbar.component';
 
 @NgModule({
   declarations: [
