--- src/app/components/card/card.component.html
@@ -1,34 +1,0 @@
-<!-- <mat-card class="example-card">
-    <mat-card-header>
-        <div mat-card-avatar class="example-header-image"></div>
-        <mat-card-title>Shiba Inu</mat-card-title>
-        <mat-card-subtitle>Dog Breed</mat-card-subtitle>
-    </mat-card-header>
-    <img mat-card-image src="https://material.angular.io/assets/img/examples/shiba2.jpg" alt="Photo of a Shiba Inu">
-    <mat-card-content>
-        <p>
-            The Shiba Inu is the smallest of the six original and distinct spitz breeds of dog from Japan.
-            A small, agile dog that copes very well with mountainous terrain, the Shiba Inu was originally
-            bred for hunting.
-        </p>
-    </mat-card-content>
-    <mat-card-actions>
-        <button mat-button>LIKE</button>
-        <button mat-button>SHARE</button>
-    </mat-card-actions>
-</mat-card>
-
-<ng-container></ng-container> -->
-<div>
-    <ul>
-        <li *ngFor="let user of users">
-            <p>{{user.id}}</p>
-            <p>{{user.firstname}}</p>
-            <p>{{user.lastname}}</p>
-            <p>{{user.email}}</p>
-            <p>{{user.phone}}</p>
-        </li>
-    </ul>
-</div>
-
-<p> Work in progress</p>
