--- src/app/components/main-tabs/main-tabs.component.html
@@ -1,8 +1,0 @@
-<mat-tab-group class='cards__wrapper'>
-    <mat-tab label="First">
-        <app-card></app-card>
-    </mat-tab>
-    <mat-tab label="Second">Content 2
-    </mat-tab>
-    <mat-tab label="Third"> Content 3 </mat-tab>
-</mat-tab-group>
