+++ src/app/users/components/card/card.component.html
@@ -1,0 +1,15 @@
+<mat-card *ngFor="let user of users" class="example-card">
+    <mat-card-header>
+        <div mat-card-avatar class="example-header-image"></div>
+        <mat-card-title>USER {{user.id}}</mat-card-title>
+        <mat-card-subtitle>{{user.firstname}} {{user.lastname}}</mat-card-subtitle>
+    </mat-card-header>
+    <img mat-card-image src="https://material.angular.io/assets/img/examples/shiba2.jpg" alt="Photo of a Shiba Inu">
+    <mat-card-content>
+        <p>{{user.email}}</p>
+        <p>{{user.phone}}</p>
+    </mat-card-content>
+</mat-card>
+
+
+<p>Loading ...</p>
