+++ src/app/users/components/main-tabs/main-tabs.component.html
@@ -1,0 +1,7 @@
+<mat-tab-group>
+    <mat-tab label="Angular lecture 1">
+        <app-card></app-card>
+    </mat-tab>
+    <mat-tab label="Angular lecture 2">Content 2
+    </mat-tab>
+</mat-tab-group>
