+++ src/app/users/components/main-tolbar/main-tolbar.component.html
@@ -1,0 +1,5 @@
+<mat-toolbar>
+    <span>Home work Angular</span>
+    <span class="example-spacer"></span>
+
+</mat-toolbar>
