+++ src/app/users/users.module.ts
@@ -1,10 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
+import { UsersPageComponent } from './pages/users-page/users-page.component';
 
 
 
 @NgModule({
-  declarations: [],
+  declarations: [
+    UsersPageComponent
+  ],
   imports: [
     CommonModule
   ]
